Move test cases to jest-free module
... so that they can be imported properly inside, e.g., a browser or the
./bin/run-example script
